+++ ouds_core/lib/components/top_bar/internal/ouds_top_bar_style_modifier.dart
@@ -1,4 +1,3 @@
-
 //
 // Software Name: OUDS Flutter
 // SPDX-FileCopyrightText: Copyright (c) Orange SA
@@ -15,13 +14,13 @@
 library;
 
 import 'dart:ui';
+
 import 'package:flutter/material.dart';
 import 'package:ouds_theme_contract/ouds_theme.dart';
 import 'package:ouds_theme_contract/theme/utilities/theme_utils.dart';
 
 /// Used to apply the right style(background, border, blur) associated to the theme mode
 class OudsTopBarStyleModifier {
-
   final BuildContext context;
 
   OudsTopBarStyleModifier(this.context);
@@ -29,10 +28,8 @@ class OudsTopBarStyleModifier {
   /// Returns the background color of the navigation bar depending on translucency and theme.
   Color getBackgroundColor(bool isTranslucent) {
     final barTheme = OudsTheme.of(context).componentsTokens(context).bar;
-    final isDark = ThemeUtils.isDarkTheme(context);
     if (isTranslucent) {
-      return isDark ? barTheme.colorBgTranslucentDark
-          : barTheme.colorBgTranslucentLight;
+      return barTheme.colorBgTranslucent;
     } else {
       return barTheme.colorBgOpaque;
     }
@@ -58,5 +55,4 @@ class OudsTopBarStyleModifier {
       sigmaY: barTheme.effectBgBlur.toDouble(),
     );
   }
-
 }